apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Maigie FastAPI backend"
authors = [
    {name = "Maigie Team"}
]
license = {text = "BUSL-1.1"}
requires-python = ">=3.11,<4.0"
dynamic = ["dependencies"]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
websockets = "^13.1"
authlib = "^1.3.0"
email-validator = "^2.1.0"
redis = {extras = ["hiredis"], version = "^5.0.0"}
celery = {extras = ["redis"], version = "^5.3.0"}
prisma = "^0.15.0"
python-json-logger = "^2.0.7"
sentry-sdk = {extras = ["fastapi"], version = "^2.0.0"}
prometheus-client = "^0.21.0"
httpx = "^0.28.1"
beautifulsoup4 = "^4.12.0"
nh3 = "^0.2.21"
trafilatura = "^2.0.0"
itsdangerous = "^2.2.0"
fastapi-mail = "^1.5.8"
bcrypt = "4.0.1"
stripe = "^10.0.0"
google-genai = "^1.30.0"
duckduckgo-search = "^7.0.0"
pypdf = "^4.0.0"
python-docx = "^1.1.0"
pinecone = "^6.0.0"
mcp = "^1.2.1"
openai = "^1.40.0"
anthropic = "^0.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.0"
black = "^25.12.0"
mypy = "^1.11.0"
coverage = "^7.12.0"
flake8 = "^7.3.0"
isort = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# --- Tool Configurations ---

# --- Tool Configurations ---

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
# F401 = Unused imports
# F841 = Unused local variables
# N818 = Exception naming
# E501 = Line length
# E203: whitespace before ':' — Black slices; keep Black as source of truth for layout
ignore = ["E501", "E203", "N818", "F401", "F841", "N803", "N815", "UP035"]

[tool.black]
line-length = 100
target-version = ["py312"]
#NEW
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"  
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:httplib2.*",
    # Third-party noise under Python 3.12+ / pytest-asyncio / FastAPI
    "ignore::DeprecationWarning:fastapi.*",
    "ignore::DeprecationWarning:starlette.*",
    "ignore::DeprecationWarning:pytest_asyncio.*",
    "ignore::DeprecationWarning:asyncio.*",
    # Prisma ships a Pydantic v1 compat shim that warns on newer Python
    "ignore:Core Pydantic V1 functionality isn't compatible:UserWarning",
    # Pydantic v2 migration (our models still use class-based Config in places)
    "ignore::pydantic.warnings.PydanticDeprecatedSince20",
]



[tool.isort]
profile = "black"

[tool.prisma]
seed = "python prisma/seed.py"